New election property "voter-privacy"
Add a new election property "voter-privacy" which disables public
display of the voter list and the dump. The default is off for the
following reasons:

The voter list is public by default because in elections with broad
qualifications, like Board elections, the biggest trust issue is the
prospect of sockpuppet accounts. Opening up the voter list helps to
assure the community that the voters are genuine users of the site.
However, this rationale may not be applicable if the qualifications are
more heavily restricted.

The dump is public by default to address voter concerns that their vote
may have been silently removed from the database. Voters are given a
receipt which they can verify against the public dump. Voters can
privately verify with keyholders that the encrypted record encodes
the expected value. And keyholders can verify that the dump they have
been given matches the one which is returned to the public for receipt
spot-checks. However, it provides information about the number of
voters, and creates a potential security issue if the key is
compromised, so it might not be appropriate in cases where the
keyholders and voters trust the integrity of the server.

Change-Id: I2b1495aa29fb0a186e27f3c75c402b9e5d2452ff