Stop cloner add/delete functions from triggering unexpectedly
HTMLFormFieldCloner uses the submit input to ocassionally pass along
values to the backend so that non-js users have access to cloner
functionality. A side effect of this is that anytime the form is
non-specifically submitted (eg. when pressing 'enter' while focused on
an input), it will also trigger the first submit input available.

On SecurePoll/create, this is a cloner add/delete button, hence the
intermittently disappearing admins.

By adding a decoy submit button, the form will submit without side
effects in those generic submit instances.

Bug: T270200
Change-Id: I246baa585245aaf9910a3ef5337f73775f47a35a