Use a validation callback for required CreatePage fields
For no-JS environments, it doesn't make sense to set certain fields
to be required:
* Fields that are conditionally shown in JS environments (these are
  not actually required in circumstances when they would be hidden)
* Fields that can be cloned (the clones are not actually required)

These need fixing ahead of the switch to OOUI (T271149), because
OOUI adds a required attribute to required inputs. This means that
the form can no longer be submitted if required fields are not
filled, so we must only require fields that are actually required.

Change-Id: I0668ee53e19a8148164ea65b17a00840cbc2fbe4