Use a validation callback for VoterEligibilityPage required fields
VoterEligibilityPage uses some HTMLFormFields that are required but
hidden conditionally depending on the value of other fields, using
JavaScript.

For no-JS environments, the fields are always visible, meaning they
always need to be filled in order for the form to submit successfully
- even when they will never be used.

This replaces the required attribute with a validation callback,
ahead of the switch to OOUI (T271150).

Change-Id: I5bfa0863de99279019fa671c9391ef4890ecc7d7