Stop deleting the temp directory after decrypting
Decrypting more than a few hundred votes results in a timeout
Part of that is a result of how long it takes to tear down/re-create
the temp directory with keys for each vote. By using a single temp
directory until the process finishes (successfully or not), we remove
that overhead and (hopefully) resolve the timeout issues. If they
continue to happen, other optimizations will have to be made (or
this will have to be converted to a job).

Since we remove the cleanup step from the decrypt function, all
pages/scripts that call it will have to clean up after themselves

Bug: T271824
Change-Id: I33de48f2994c3b516de14300e970109e6e47aaf6