Allow decrption key to be entered via tally page
Instead of requiring that a GPG decryption key be added to the
database before an election can be tallied, allow one to be added via
the TallyPage form.

Follow the existing pattern of abstracting out particular encryption
types using the Crypt class. Crypt::getTallyDescriptors allows fields
to be added to the TallyPage forms. Crypt::updateTallyContext allows
form data to be added to the election context.

Update GpgCrypt to accept and use a decryption key if there is none
stored in the database. For existing elections with keys in the
database, continue to use the stored keys.

Bug: T271830
Change-Id: I67e1cd86ae226e1deeab9f9ee7fed4857e6c07bb