VoterEligibilityPage: Check date validity in a validation callback
The date callback was a part of the processConfig() which is
run no matter what. Even if "list_edits-between" wasn't
an eligibility requirement, the date passed through had to be
valid (after the start date which also meant that start date
had to exist).

This patch moves that function out of processConfig(),
allowing for the eligibility to be saved without setting
the start and end date for "list_edits-between"

Bug: T180762
Change-Id: I6807eea88fac387aa311b46e6fb51fbd045aaeb2