VoterEligibilityPage: Manually validate dynamically required fields
In no-js, submitting an invalid eligibility requirement will
result in showing all invalid eligibility requirements, even for
eligibility requirements not being used.

This also happens with js enabled but because the fields are
dynamically shown/hidden, this is less immediately obvious.

This is purely aesthetic, as HTMLForm::trySubmit() will ignore
all hidden fields (determined by isHidden()) so it doesn't
matter if some required fields are not filled out, so long as
they're considered hidden.

This patch manually sets input requirements, taking into account
pre-requsite checks. For example, for the 'central-block-threshold'
to be necessary (and for input requirements to matter), its parent
'not-centrally-blocked' needs to be selected first.

Bug: T180762
Change-Id: Ic8a69a2b73126fd0901ccd6e67dc567e05686f0e