Combine the two TallyPage forms into one form
If an election is encrypted, show one form with an optional file
input, rather than two forms (one with no file input and one with a
mandatory file input).

Remove the extra handling for two forms, and associated messages.

Preserve the existing functionality:
* If an election is not encrypted, only have a submit button
* If an election is encrypted, but additional info is needed to
  decrypt it, add extra fields depending on the encryption method

Finally, remove the help text at the top of the form and replace it
with some help text for the file input.

Bug: T272916
Change-Id: I1e184dcccc9fd815c307aa090e04e7bd73d6521e