Get rid of resources/SecurePoll.js
Delete the ballot error code entirely -- it's been broken for a long
time because it tries to call the setup function before the JS is
output at the end of the page. CSS seems to have been added to
replicate what it'd have been doing, so removing it rather than fixing
it seems reasonable.

Move the admin vote strike code into the existing resourceloader module.
I'll refactor it to not need globals as part of a rewrite of that form.

Change-Id: I3a65d0d6faf462595f06dbcae84bfc56825cdacd