Tally elections via DeferredUpdate to avoid timeout
Perform the tally after the request, to avoid timing out and
requiring the user to run a maintenance script.

Changes to the tally page:
* If the last tally errored, show the error on the TallyPage
* While a tally is in progress, add a warning to the TallyPage
  and disable the form
* Once a tally has finished, show the results on the TallyPage
* Keep showing the form, to allow re-tallying

The deferred update can't be used for elections that are uploaded
via an XML dump since it is not clear where to return to for the
results. The possibility of a timeout error therefore remains, if
a user uploads a large election. However, this feature is rarely
used and may be removed.

Bug: T277222
Bug: T277224
Change-Id: I4bedcde13838fa29bec496ed54b6e4dc3ca82a7d