Add scripts for 2021 voter qualification
I updated the way that user information is queried. The idea of the
original code is that the user cache speeds up isAnon(), and
User::getId() is fast in any case. But bitrot meant that in the worst
case it was doing two queries: one from User::loadFromDatabase() and
another from findActorIdInternal(). Maybe a join would be better, but
this way is simple and likely to work for another couple of years.

Change-Id: I9d9e5da022f41fcc70785af67625364d2ccefbdb
(cherry picked from commit f89c0772191e10a5b33be2dee029d1ded17b9c8b)