Disallow elimination of already elected candidates
Elected candidates should never be considered for elimination.

In an edge case, it may be possible for an elected candidate's
vote count to be considered the "lowest" in the ranking because
the transfer of surplus to a hopeful candidate increases their
vote count. This would cause the elected candidate to be declared
eliminated.

Bug: T290001
Change-Id: I0486d40e742a609ff2d6e144d12678fa24f0395e