Improve dependency injection into Ballot classes
Remove all RequestContext access in the Ballot classes. Instead have
VotePage inject the required state into the Ballot.

Also fix some code style issues in RadioRangeBallot.
Ballot::submitForm() assumes the return value of submitQuestion() is a
string even if the Status is an error. If there is an error, the string
is thrown away.

Change-Id: Ic4fe7ffd0d3b47a3ab4690cd3285abb2bd8c00d6