Fix filepath for 1.39 postgres SQL patch
Fix the filepath specified in InstallHandler.php to the SQL file
patch-securepoll_msgs-drop-pk.sql for the postgres DB type. The
current filepath is incorrect and causes install.php to fail when
using postgres.

This also therefore causes the experimental postgres tests to fail
if SecurePoll is included in the zuul clone.

QA steps:
* Running the "experimental" checks should find that the install
when using the postgres DB now succeeds without error.

Bug: T339829
Change-Id: Ic3422e5a43b08e173ae852078963bf7758bcbdb0