Clear UserGetRights hook handlers in SpecialSecurePollLogTest
The test doesn't need them, and the one in CentralAuth can trigger DB
access, which would make this test fail because it's not in the Database
group.

Also add SecurePollLogPagerTest to the Database group:
IndexPager::__construct obtains a DB connection, and all pager classes
generally expect the DB to be available.

Bug: T345372
Change-Id: I5fc5b026fd13a394f255ab8c5bfb91331b47557c