Guard against loading unecessary JS in contexts that dont need it
Currently the ext.securepoll.htmlform was loading the code for
translating pages, voting and listing all at the same time.
However this led to cases where the JS would try to execute in
unexpected ways (for example trying to find and infuse a
translation button on a voting/listing page etc).

This patch adds a config variable 'SecurePollSubPage' which
is then used to only load the required JS for a specific page.

In addition to this, the JS on a voting page was incorrectly
identifying dropdown elements and trying to infuse them. This
has also been guarded against in this patch.

Bug: T352175
Change-Id: Ie4468a2f5421a4797ce09275f77494f09fc912b6