Add encryption method using openssl
This relies on the openssl extension being available in PHP and is
currently hardcoded to only make use of RSA keys for ease of initial
implementation and to reduce the amount of code for initial review.
Later commits may wish to expand upon the list of allowable encryption
and signing algorithms.

The functionality of the file largely matches that of the GpgCrypt
class. The receipts are generated to be signed JWTs as there are plenty
of tools (both online and offline) capable of validating JWTs so that a
user can have some assurance they hold a valid vote. The public portion
of the signing key is not made available however, and would need to be
done externally.

Tests were added to ensure that the encryption roundtrips successfully
and that various "invalid" formations are correctly disregarded.

Bug: T209892
Change-Id: Iaab9f69bba835eab5d3f5a8585ae17b230169498