tests: Replace workaround that concatenated an internal file
Follows-up c42f627979 (I3d3ac44de3). This was somewhat confusing as it
re-assigned the test module's `module.exports` object twice and
ignored it both times, but was working due to defining it in the local
scope.

Import the private file directly instead. Another way, if there is no
convenient way to export it directly in an existing code base, is to
attach it to a `/** @private */` property on the mw object like
`mw.securePoll.something`.

Change-Id: I8fcf8d46032a199c2638d9506b20318f1042ffcd