Don't disable write actions for SecurePoll pages
Disabling action=edit causes the "View source" link to lead to "No
such action". There's no harm in allowing the JSON source to be seen.
All we need to do is prevent editing – which is already being blocked
by the TitleQuickPermissions hook.

Disabling action=info causes "Page information" to break similarly. This
is not a write action and works fine for SecurePoll pages.

Don't disable the fictitious action=revert, because it doesn't exist in
the first place.

Don't disable action=delete, protect and unprotect either, because –
again due to the TitleQuickPermissions hook – these options are not
shown in the UI. If a user constructs the link anyway, they see the
appropriate error message that SecurePoll namespace is read-only.

Bug: T264476
Change-Id: If95ec88968e471febe17562075a9f8b4a50a74d6