Use fixed-point arithmetic for the STV algorithm
This fixes an issue where the tallier would get stuck in an infinite
loop due to floating-point errors accumulating on the surplus over
multiple rounds.

Test data was also updated in 'tests/phpunit/unit/fixtures' since the
tests check if the calculations for the rounds have changed. The bc math
functions use strings instead of floats causing these tests to break, so
new snapshots of the 'rounds' array were added to the tests to fix this.
Also less rounds are needed to complete elections when using fixed-point
arithmetic as well.

Bug: T291821
Bug: T289185
Bug: T361077
Change-Id: Iaa1f2dce6be5848d0eeaf0a0ca55ddd386f68bb6