Special:SecurePoll: add status field to pager
This will make it easier to tell if an election is not started,
in progress, or completed, without having to pay
attention to the start date and without having to do
calculations in your head.

This will also make things make more sense when we start
disabling certain links depending on status.

Bug: T392446
Change-Id: I480dc6c5776f2d4081d512ab7a02326f22380ec2