MainElectionsPager: don't hyperlink disabled pages
Depending on the status of a particular poll/election, visiting
a page such as Special:SecurePoll/tally may just give you an
error message such as "Sorry, you cannot tally the election
until after voting is complete."

So let's not hyperlink these pages. No use making the reader
guess which pages are enabled and which are disabled.

Bug: T392446
Change-Id: I3ec525cb73c64cf3a078c1ba07a7f7581495082c