Handle missing 'wiki' parameter
If the 'wiki' parameter was null, we were passing `null` to the
SiteConfiguration, which doesn't expect this.  Skip the call to
SiteConfiguration::siteFromDB() and coerce to the empty string
for the call to SiteConfiguration::get().

I'm not sure this is the *correct* behavior, but it should match
the *existing* behavior, at least.  It's possible we should just
return Status::newFatal() in this case.

Bug: T391871
Change-Id: I98d22b018201ad2bafcddb033226586b67bfee8f