Prevent saving of virtual ids to namespace
This reverts the hacks added in I3829b327b. The entity information built
from the form data initially contains virtual ids (-1, -2, and so on)
for new entities. As and when the real ids of these entities become
known (after the entity is inserted to the db), replace the virtual id
with the real id in all the data structures in FormStore. This
eventually gets used while saving to the namespace.

Bug: T397482
Change-Id: Id739845a4e8f1b7dd577910689142cb821fd36df