Streamline group-related eligibility options
Why?
* The 'allow-usergroups' option allows configuring groups that are
always allowed to vote, regardless of all other options or lists. Even
blocked users can vote even if "Must not be sitewide blocked" is
checked. This is not clear enough, since the label just says
"regardless of edits or other groups".
* The 'need-group' eligibility option currently exists but is not
exposed in UI. It limits voting to users in the specified group, if
they meet the other requirements as well.

What?
* Expose the 'need-group' option to UI, and allow multiple groups (voter
needs to be in any one of them). This is backward-compatible because
like other multiselect options, a '|'-separated string is internally
used for multiple values.
* Move down the 'allow-usergroups' option to be just under "Override
list". This emphasises the similar override behaviour of this option.
Also adjust the label.

Bug: T397596
Change-Id: If645d88c16e6ea8f04dbcb6a2675abf78fcb4927