VoterEligibilityPage: Fix writing eligibility lists to namespace
This fixes logging of lists to namespace always being one edit out of
date.

The usual strategy of using the form data for writing to namespace
(instead of reading it from the database) doesn't work here due to the
use of $dbw->insertSelect() which inserts new rows based on what's
already in the database.

So, we avoid a stale read by reading back the data *before* the
transaction is committed. This guarantees that all data written will
be read back.

Bug: T397592
Change-Id: Ib76250c5f3872020a4aad86b5d5c9e422fb89dae