VoterEligibilityPage: Simplify validation of required and min values
For form fields which are not always visible, validations are done
through custom validation-callbacks. (This is because in no-JS mode,
all fields are always visible, and HTML5 validations may fail on fields
which are not relevant, eg. the central block threshold input when
"Must not be blocked on too many sites" is unchecked.)

Converted the validation callbacks to anonymous functions. Also,
moved the min validations out from them as they don't need to be in
the callbacks -  browsers don't do min/max validations on empty fields.
Tested across Chrome and Firefox in no-JS mode.

Change-Id: I93ee54756379dc6199e2080b579fbce922fa6d43