Add eligibility options to require minimum edit count in time intervals
Introduce two new eligibility options:
 * Minimum edits before a given date
 * Minimum edits between two given dates

Eligibility criteria based on these options are common for many
elections.

The checks are done when the user attempts to vote, using fast db
queries that join on indexed columns. Instead of using COUNT(*) on
revision table which can become very slow for users with millions of
edits (T302761), this does a SELECT 1 with a LIMIT equal to the number
of edits to check. We limit the maximum allowed value to 1000 for now,
since checking for a very large number of edits can lead to the same
problem.

This is intended to supersede automatic list generation, which can be
removed in a future commit (Iabd4c9b).

Bug: T397565
Change-Id: Idc9ef76b36aadc122ef80a2b4f449f7fb0a59d82