Remove automatic list generation
Equivalent eligibility options that are instead checked at the time of
voting are added in Idc9ef76b.

Automatic list generation is slow and unreliable. Enwiki has 40 million
user accounts. Checking each one and saving the list of eligible
accounts can take days. This feature does not have any current usage,
and is prone to issues if it starts getting used. For instance, the
query to find users with greater than N edits before a date uses
COUNT(*) on revision table which is unlikely to work as many users have
massive edit counts (same issue as T302761).

The setup interface is also buggy - the options to include and exclude
user groups don't always show up. Including groups is already covered
under basic options. Excluding isn't, but the only practical use case
for that is to exclude bots, which is covered.

Bug: T397565
Change-Id: Iabd4c9b834d7ba2fa6919dedfcfd93f542729e33