fix exception when editing or deleting a voter list
This exception only occurs when $wgSecurePollUseNamespace =
false; and $wgSecurePollUseMediaWikiNamespace = true; It was
because the page title was incorrect. Specifically, it was trying to
write to a non-existent namespace.

Testing notes: Keep in mind T397592 (Outdated values of eligibility
settings and lists saved to namespace) when testing. You may need to
save twice to get MediaWiki to write a revision.

Bug: T398680
Change-Id: I66343acbea60566a01c6b648917c8915ce9335b5