Fix logging to MediaWiki namespace if subpages are disabled
With subpages disabled in MediaWiki namespace (as is the case on
enwiki), $title->getRootText() === 'SecurePoll' does not evaluate to
true on titles of the form MediaWiki:SecurePoll/*.

Since SecurePollContentHandler allows creation of pages with that
content model only on titles that pass the check, the pages didn't get
created. Also, the protection from TitleQuickPermissions hook wasn't
getting applied as it was using the same check.

Bug: T398716
Change-Id: If0b0891060c14f44ee81e2323d4ca7bb946828a9